Updating formatting and fixing compiler warnings

+++ vtkm/cont/testing/UnitTestDataSetBuilderUniform.cxx
@@ -188,13 +188,13 @@ UniformTests()
     std::cout << "3D case" << std::endl;
     vtkm::Id numPoints = dimensions[0]*dimensions[1]*dimensions[2];
     vtkm::Id numCells = (dimensions[0]-1)*(dimensions[1]-1)*(dimensions[2]-1);
-    std::vector<T> varP3D(numPoints);
-    for (int i = 0; i < numPoints; i++)
+    std::vector<T> varP3D(static_cast<unsigned long>(numPoints));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numPoints); i++)
     {
       varP3D[i] = static_cast<T>(i*1.1f);
     }
-    std::vector<T> varC3D(numCells);
-    for (int i = 0; i < numCells; i++)
+    std::vector<T> varC3D(static_cast<unsigned long>(numCells));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numCells); i++)
     {
       varC3D[i] = static_cast<T>(i*1.1f);
     }
@@ -207,13 +207,13 @@ UniformTests()
     numPoints = dimensions[0]*dimensions[1];
     numCells = (dimensions[0]-1)*(dimensions[1]-1);
     bounds.Z = vtkm::Range(0, 0);
-    std::vector<T> varP2D(numPoints);
-    for (int i = 0; i < numPoints; i++)
+    std::vector<T> varP2D(static_cast<unsigned long>(numPoints));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numPoints); i++)
     {
       varP2D[i] = static_cast<T>(i*1.1f);
     }
-    std::vector<T> varC2D(numCells);
-    for (int i = 0; i < numCells; i++)
+    std::vector<T> varC2D(static_cast<unsigned long>(numCells));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numCells); i++)
     {
       varC2D[i] = static_cast<T>(i*1.1f);
     }
@@ -229,13 +229,13 @@ UniformTests()
     numCells = dimensions[0]-1;
     bounds.Y = vtkm::Range(0, 0);
     bounds.Z = vtkm::Range(0, 0);
-    std::vector<T> varP1D(numPoints);
-    for (int i = 0; i < numPoints; i++)
+    std::vector<T> varP1D(static_cast<unsigned long>(numPoints));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numPoints); i++)
     {
       varP1D[i] = static_cast<T>(i*1.1f);
     }
-    std::vector<T> varC1D(numCells);
-    for (int i = 0; i < numCells; i++)
+    std::vector<T> varC1D(static_cast<unsigned long>(numCells));
+    for (unsigned long i = 0; i < static_cast<unsigned long>(numCells); i++)
     {
       varC1D[i] = static_cast<T>(i*1.1f);
     }
